add `kr codesign test`
